add hcd_devtree_get_info()
remove usbh_hcd.h
