+++ src/host/hcd_attr.h
@@ -35,6 +35,7 @@
 
 //------------- NXP -------------//
 #if TU_CHECK_MCU(LPC175X_6X) || TU_CHECK_MCU(LPC177X_8X) || TU_CHECK_MCU(LPC40XX)
+  #define HCD_ATTR_OHCI
 
 #elif TU_CHECK_MCU(LPC18XX) || TU_CHECK_MCU(LPC43XX)
   #define HCD_ATTR_EHCI_TRANSDIMENSION
